requirements.txt
# OCR正規化パイプライン 依存パッケージ
# インストール: pip install -r requirements.txt

# 画像処理
Pillow>=10.0.0

# OCR
pytesseract>=0.3.10

# PDF処理
PyMuPDF>=1.23.0

# Excel/CSV出力
openpyxl>=3.1.2

# 注意:
# - Tesseract OCRエンジンは別途システムにインストールする必要があります
#   Ubuntu/Debian: sudo apt-get install tesseract-ocr tesseract-ocr-jpn tesseract-ocr-kor
#   macOS: brew install tesseract tesseract-lang
#   Windows: https://github.com/UB-Mannheim/tesseract/wiki からダウンロード
